+++ core/src/main/java/org/seedstack/seed/core/internal/crypto/CryptoPlugin.java
@@ -5,6 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.seed.core.internal.crypto;
 
 import com.google.common.base.Strings;
@@ -17,9 +18,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import javax.net.ssl.KeyManager;
 import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManager;
+import javax.net.ssl.X509KeyManager;
 import org.seedstack.seed.SeedException;
 import org.seedstack.seed.core.internal.AbstractSeedPlugin;
 import org.seedstack.seed.crypto.CryptoConfig;
@@ -30,15 +30,12 @@ import org.slf4j.LoggerFactory;
 
 public class CryptoPlugin extends AbstractSeedPlugin implements SSLProvider {
     private static final Logger LOGGER = LoggerFactory.getLogger(CryptoPlugin.class);
-    private static final String ALIAS = "alias";
-
     private final Map<Key<EncryptionService>, EncryptionService> encryptionServices = new HashMap<>();
     private final Map<String, KeyStore> keyStores = new HashMap<>();
-    private final Map<String, CryptoConfig.KeyStoreConfig> keyStoreConfigs = new HashMap<>();
-    private final List<KeyPairConfig> keyPairConfigs = new ArrayList<>();
-
-    private CryptoConfig.SSLConfig sslConfig;
+    private KeyStore trustStore;
     private SSLContext sslContext;
+    private List<KeyManagerAdapter> keyManagerAdapters;
+    private Class<? extends X509KeyManager> keyManagerClass;
 
     @Override
     public String name() {
@@ -46,76 +43,62 @@ public class CryptoPlugin extends AbstractSeedPlugin implements SSLProvider {
     }
 
     @Override
-    public Object nativeUnitModule() {
-        return new CryptoModule(encryptionServices, keyStores, sslContext);
-    }
-
-    @Override
     public InitState initialize(InitContext initContext) {
         CryptoConfig cryptoConfig = getConfiguration(CryptoConfig.class);
-
-        // Load key stores
         KeyStoreLoader keyStoreLoader = new KeyStoreLoader();
-        for (Map.Entry<String, CryptoConfig.KeyStoreConfig> entry : cryptoConfig.keyStores().entrySet()) {
-            keyStores.put(entry.getKey(), keyStoreLoader.load(entry.getKey(), entry.getValue()));
-            keyStoreConfigs.put(entry.getKey(), entry.getValue());
+        this.trustStore = loadTrustStore(cryptoConfig, keyStoreLoader);
+        this.keyStores.putAll(loadKeyStores(cryptoConfig, keyStoreLoader));
+        this.encryptionServices.putAll(configEncryptionServices(cryptoConfig));
+        SSLBuilder sslBuilder = new SSLBuilder(trustStore, keyStores);
+        this.sslContext = sslBuilder.getSSLContext(cryptoConfig);
+        this.keyManagerAdapters = sslBuilder.getKeyManagerAdapters();
+        this.keyManagerClass = cryptoConfig.ssl().getX509KeyManager();
+        return InitState.INITIALIZED;
+    }
+
+    private KeyStore loadTrustStore(CryptoConfig cryptoConfig, KeyStoreLoader keyStoreLoader) {
+        CryptoConfig.StoreConfig trustStoreConfig = cryptoConfig.getTrustStore();
+        if (trustStoreConfig == null) {
+            return null;
+        } else {
+            LOGGER.info("Loading truststore from {}", trustStoreConfig.getPath());
+            return keyStoreLoader.load("<truststore>", trustStoreConfig);
         }
+    }
+
+    private Map<String, KeyStore> loadKeyStores(CryptoConfig cryptoConfig, KeyStoreLoader keyStoreLoader) {
+        Map<String, KeyStore> loadedKeyStores = new HashMap<>();
+        cryptoConfig.keyStores().entrySet().stream()
+                .peek(e -> LOGGER.info("Loading keystore '{}' from {}", e.getKey(), e.getValue().getPath()))
+                .forEach(e -> loadedKeyStores.put(e.getKey(), keyStoreLoader.load(e.getKey(), e.getValue())));
+        return loadedKeyStores;
+    }
+
+    private Map<Key<EncryptionService>, EncryptionService> configEncryptionServices(CryptoConfig cryptoConfig) {
+        List<KeyPairConfig> keyPairConfigs = new ArrayList<>();
 
         // Retrieve key pair configurations
         KeyPairConfigFactory keyPairConfigFactory = new KeyPairConfigFactory(cryptoConfig);
         for (Map.Entry<String, KeyStore> entry : keyStores.entrySet()) {
-            this.keyPairConfigs.addAll(keyPairConfigFactory.create(entry.getKey(), entry.getValue()));
+            keyPairConfigFactory.create(entry.getKey(), entry.getValue()).stream()
+                    .peek(keyPairConfig -> LOGGER.debug("Encryption service '{}' defined from keystore '{}'",
+                            keyPairConfig.getAlias(),
+                            entry.getKey()))
+                    .forEach(keyPairConfigs::add);
         }
 
-        // Prepare encryption service bindings
-        encryptionServices.putAll(new EncryptionServiceBindingFactory().createBindings(cryptoConfig,
-                this.keyPairConfigs,
-                keyStores));
-        LOGGER.debug("Registered {} cryptographic key(s)", encryptionServices.size());
-
-        // init SSL context if possible
-        this.sslConfig = cryptoConfig.ssl();
-        this.sslContext = configureSSL(this.sslConfig).orElse(null);
-
-        return InitState.INITIALIZED;
+        // Create encryption service bindings
+        return new EncryptionServiceBindingFactory().createBindings(keyPairConfigs, keyStores);
     }
 
-    private Optional<SSLContext> configureSSL(CryptoConfig.SSLConfig sslConfig) {
-        SSLLoader sslLoader = new SSLLoader();
-
-        TrustManager[] trustManagers;
-        KeyStore trustStore = keyStores.get(sslConfig.getTrustStore());
-        if (trustStore == null) {
-            return Optional.empty();
-        } else {
-            trustManagers = sslLoader.getTrustManager(trustStore);
-        }
-
-        return configureKeyManagers(sslConfig).map(
-                keyManagers1 -> sslLoader.getSSLContext(this.sslConfig.getProtocol(), keyManagers1, trustManagers));
-    }
-
-    private Optional<KeyManager[]> configureKeyManagers(CryptoConfig.SSLConfig sslConfig) {
-        String keyStoreName = sslConfig.getKeyStore();
-        CryptoConfig.KeyStoreConfig keyStoreConfig = keyStoreConfigs.get(keyStoreName);
-
-        if (keyStoreConfig == null) {
-            return Optional.empty();
-        }
-
-        String aliasName = sslConfig.getAlias();
-        CryptoConfig.KeyStoreConfig.AliasConfig aliasConfig = keyStoreConfig.getAliases().get(aliasName);
-        if (aliasConfig == null) {
-            return Optional.empty();
-        }
-        if (Strings.isNullOrEmpty(aliasConfig.getPassword())) {
-            throw SeedException.createNew(CryptoErrorCode.MISSING_ALIAS_PASSWORD)
-                    .put(ALIAS, aliasName)
-                    .put("ksName", keyStoreName);
-        }
-
-        return Optional.of(
-                new SSLLoader().getKeyManagers(keyStores.get(keyStoreName), aliasConfig.getPassword().toCharArray()));
+    @Override
+    public Object nativeUnitModule() {
+        return new CryptoModule(encryptionServices,
+                keyStores,
+                trustStore,
+                sslContext,
+                keyManagerAdapters,
+                keyManagerClass);
     }
 
     @Override
@@ -125,6 +108,16 @@ public class CryptoPlugin extends AbstractSeedPlugin implements SSLProvider {
 
     @Override
     public CryptoConfig.SSLConfig sslConfig() {
-        return sslConfig;
+        return getConfiguration(CryptoConfig.SSLConfig.class);
+    }
+
+    static EncryptionService getMasterEncryptionService(EncryptionServiceFactory encryptionServiceFactory,
+            CryptoConfig.KeyStoreConfig masterKeyStoreConfig, String alias) {
+        CryptoConfig.KeyStoreConfig.AliasConfig aliasConfig = masterKeyStoreConfig.getAliases().get(alias);
+        if (aliasConfig == null || Strings.isNullOrEmpty(aliasConfig.getPassword())) {
+            throw SeedException.createNew(CryptoErrorCode.MISSING_MASTER_KEY_PASSWORD);
+        }
+        return encryptionServiceFactory.create(alias,
+                aliasConfig.getPassword().toCharArray());
     }
 }