+++ core/src/main/java/org/seedstack/seed/core/internal/crypto/KeyPairConfig.java
@@ -5,20 +5,19 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.seed.core.internal.crypto;
 
 class KeyPairConfig {
     private final String alias;
     private final String password;
-    private final String certificateLocation;
     private final String qualifier;
     private final String keyStoreName;
 
-    KeyPairConfig(String keyStoreName, String alias, String password, String certificateLocation, String qualifier) {
+    KeyPairConfig(String keyStoreName, String alias, String password, String qualifier) {
         this.keyStoreName = keyStoreName;
         this.alias = alias;
         this.password = password;
-        this.certificateLocation = certificateLocation;
         this.qualifier = qualifier;
     }
 
@@ -28,10 +27,6 @@ class KeyPairConfig {
 
     String getPassword() {
         return password;
-    }
-
-    String getCertificateLocation() {
-        return certificateLocation;
     }
 
     String getQualifier() {