+++ core/src/main/java/org/seedstack/seed/core/internal/lifecycle/LifecyclePlugin.java
@@ -54,10 +54,9 @@ public class LifecyclePlugin extends AbstractSeedPlugin implements LifecycleMana
         for (Class<?> candidate : initContext.scannedSubTypesByParentClass().get(LifecycleListener.class)) {
             if (LifecycleListener.class.isAssignableFrom(candidate)) {
                 lifecycleListenerClasses.add((Class<? extends LifecycleListener>) candidate);
-                LOGGER.trace("Detected lifecycle listener {}", candidate.getCanonicalName());
+                LOGGER.debug("Lifecycle listener {} detected", candidate.getCanonicalName());
             }
         }
-        LOGGER.debug("Detected {} lifecycle listener(s)", lifecycleListenerClasses.size());
         return InitState.INITIALIZED;
     }
 
@@ -73,7 +72,7 @@ public class LifecyclePlugin extends AbstractSeedPlugin implements LifecycleMana
 
         for (LifecycleListener lifecycleListener : sortedLifecycleListeners) {
             try {
-                LOGGER.info("Executing started method of lifecycle listener {}",
+                LOGGER.info("Executing started() method of lifecycle listener {}",
                         lifecycleListener.getClass().getName());
                 lifecycleListener.started();
             } catch (Exception e) {
@@ -102,7 +101,7 @@ public class LifecyclePlugin extends AbstractSeedPlugin implements LifecycleMana
 
         for (LifecycleListener lifecycleListener : sortedLifecycleListeners) {
             try {
-                LOGGER.info("Executing stopping method of lifecycle listener {}",
+                LOGGER.info("Executing stopping() method of lifecycle listener {}",
                         lifecycleListener.getClass().getName());
                 lifecycleListener.stopping();
             } catch (Exception e) {
@@ -118,12 +117,11 @@ public class LifecyclePlugin extends AbstractSeedPlugin implements LifecycleMana
             Method closeMethod = autoCloseable.getClass().getMethod("close");
             if (!Annotations.on(closeMethod).traversingOverriddenMembers().find(Ignore.class).isPresent()) {
                 if (autoCloseableObjects.add(autoCloseable)) {
-                    LOGGER.info("Registered auto-closeable {} for closing at shutdown",
+                    LOGGER.info("Auto-closeable {} registered for closing at shutdown",
                             autoCloseable.getClass().getName());
                 }
             } else {
-                LOGGER.debug("Ignored registration of auto-closeable {} for closing at shutdown",
-                        autoCloseable.getClass().getName());
+                LOGGER.debug("Ignored registration of auto-closeable {}", autoCloseable.getClass().getName());
             }
         } catch (NoSuchMethodException e) {
             throw SeedException.wrap(e, CoreErrorCode.UNEXPECTED_EXCEPTION);