+++ specs/src/main/java/org/seedstack/seed/crypto/spi/SSLProvider.java
@@ -5,6 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.seed.crypto.spi;
 
 import io.nuun.kernel.api.annotations.Facet;
@@ -21,7 +22,7 @@ public interface SSLProvider {
     /**
      * Provides an {@link javax.net.ssl.SSLContext} configured during the init phase.
      *
-     * @return an SSL context, or null before the init phase
+     * @return an optional SSL context, being empty if no SSLContext could be constructed.
      */
     Optional<SSLContext> sslContext();