+++ web/undertow/src/main/java/org/seedstack/seed/undertow/internal/ServerFactory.java
@@ -5,6 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.seed.undertow.internal;
 
 import io.undertow.Handlers;
@@ -12,8 +13,6 @@ import io.undertow.Undertow;
 import io.undertow.UndertowOptions;
 import io.undertow.server.HttpHandler;
 import io.undertow.server.handlers.PathHandler;
-import java.util.Optional;
-import javax.net.ssl.SSLContext;
 import org.seedstack.seed.SeedException;
 import org.seedstack.seed.crypto.CryptoConfig;
 import org.seedstack.seed.crypto.spi.SSLProvider;
@@ -41,37 +40,37 @@ class ServerFactory {
 
         Undertow.Builder builder = Undertow.builder().setWorker(xnioWorker);
 
-        if (serverConfig.isHttps()) {
-            builder = configureHttps(builder, sslProvider);
+        if (!serverConfig.isHttp() && !serverConfig.isHttps()) {
+            throw SeedException.createNew(UndertowErrorCode.NO_LISTENER_CONFIGURED);
         } else {
-            builder = configureHttp(builder);
-        }
-
-        if (serverConfig.isHttp2()) {
-            LOGGER.info("HTTP/2 support is enabled");
-            builder.setServerOption(UndertowOptions.ENABLE_HTTP2, serverConfig.isHttp2());
+            if (serverConfig.isHttp()) {
+                builder = configureHttp(builder);
+            }
+            if (serverConfig.isHttps()) {
+                builder = configureHttps(builder, sslProvider);
+                if (serverConfig.isHttp2()) {
+                    LOGGER.info("HTTP/2 support is enabled");
+                    builder.setServerOption(UndertowOptions.ENABLE_HTTP2, serverConfig.isHttp2());
+                }
+            }
         }
 
         return builder.setHandler(path).build();
     }
 
     private Undertow.Builder configureHttp(Undertow.Builder builder) {
+        LOGGER.info("Undertow listening for HTTP on {}:{}", serverConfig.getHost(), serverConfig.getPort());
         return builder.addHttpListener(serverConfig.getPort(), serverConfig.getHost());
     }
 
     private Undertow.Builder configureHttps(Undertow.Builder builder, SSLProvider sslProvider) {
-        Optional<SSLContext> sslContext = sslProvider.sslContext();
+        LOGGER.info("Undertow listening for HTTPS on {}:{}", serverConfig.getHost(), serverConfig.getSecurePort());
         CryptoConfig.SSLConfig sslConfig = sslProvider.sslConfig();
-        if (sslContext.isPresent()) {
-            return builder
-                    .addHttpsListener(serverConfig.getPort(), serverConfig.getHost(), sslContext.get())
-                    .setSocketOption(Options.SSL_CLIENT_AUTH_MODE,
-                            SslClientAuthMode.valueOf(sslConfig.getClientAuthMode().toString()));
-        } else {
-            throw SeedException.createNew(UndertowErrorCode.MISSING_SSL_CONTEXT)
-                    .put("ksName", sslConfig.getKeyStore())
-                    .put("tsName", sslConfig.getTrustStore())
-                    .put("alias", sslConfig.getAlias());
-        }
+        return builder
+                .addHttpsListener(serverConfig.getSecurePort(), serverConfig.getHost(), sslProvider.sslContext()
+                        .orElseThrow(() -> SeedException.createNew(UndertowErrorCode.MISSING_SSL_CONTEXT)
+                                .put("ksName", sslConfig.getKeystore())))
+                .setSocketOption(Options.SSL_CLIENT_AUTH_MODE,
+                        SslClientAuthMode.valueOf(sslConfig.getClientAuthMode().toString()));
     }
 }